Setup IgContent Composable
